+++ lib/routes/Character/components/CharacterDialog.tsx
@@ -28,7 +28,7 @@ import React, { useState } from "react";
 import { Prompt } from "react-router";
 import { ContentEditable } from "../../../components/ContentEditable/ContentEditable";
 import { FateLabel } from "../../../components/FateLabel/FateLabel";
-import { ICharacter } from "../../../contexts/CharactersContext";
+import { ICharacter } from "../../../contexts/CharactersContext/CharactersContext";
 import { useTextColors } from "../../../hooks/useTextColors/useTextColors";
 import { useTranslate } from "../../../hooks/useTranslate/useTranslate";
 import { useCharacter } from "../hooks/useCharacter";
